Custom_apps/appli-web-postgresql/swagger/requirements.txt
Flask
psycopg2-binary
flasgger
